Added right triangle support, fixed GetHeight().

+++ Shapes.Triangle/TriangleAreaExtension.cs
@@ -5,7 +5,7 @@ public static class TriangleAreaExtension
     public static double GetArea(this Triangle triangle)
     {
         var height = GetHeight(triangle);
-        var result = (triangle.SideA * height) / 2;
+        var result = (triangle.SideA * height) / (double)2;
 
         return result;
 
@@ -15,7 +15,7 @@ public static class TriangleAreaExtension
         var p = GetHalfPerimeter(triangle);
 
         var sqrtParam = p * ((p - triangle.SideA) * (p - triangle.SideB) * (p - triangle.SideC));
-        var result = 2 / triangle.SideA * Math.Sqrt(sqrtParam);
+        var result = (double)2 / triangle.SideA * Math.Sqrt(sqrtParam);
 
         return result;
     }
